Rest of code for #2277 - Deleted errands - different kinds (the first part was in the previous commit I made).
Added a new report called Deleted Errands that will is like the classification
report only it will show classifications given to deleted errands.

Also added an option to the replies report to decide if you want it to
include deleted errands and if you only want to show deleted errands with
a specific classification.

This commit also includes some fixes to Reports.