Fixed Closure Status report
Since each key is fetched with a different sql query the results may have
different lengths. This means we can't simply iterate over the returned data
to create the table. Instead we need to iterate over the keys and test for
each row if there is data for that key otherwise we print '0'.