Fix problem with sorting in reports
Weekdays and months were not sorted in the right order.
Weekdays should be sorted from Monday to Sunday and months from
January to December.

The reason weekdays and months were not sorted in the right order
was because how the output is pre-built before values are fetched
from the database.
The idea with pre-building a report is that if you run a report
from January 1 to December 31 and you were you want to see
the output per month then you previously would only see the months
you have data for in the database.
By pre-building the reports and filling in 0 for each month then
we make sure all months between January 1 and December 31 will be
show in the report.

They way the system figures out what values should be in the pre-built
report is by going through from the start time to the end time and see
what months are in between.
This becomes a problem when you want to look at a report from 2013 July
to 2014 June. In that report the months should still be sorted from
January to December but because we started from 2013 July the pre-build
code would first see July and last see June.

To solve this the pre-built values are now sorted. To make sure
they are sorted in the right order we look at the numeric representation
for months and weekdays.