Fix year calculation bug for drilldown
The original code uses .endTime() which when passed to localYearFromGMTtimestamp() will
get adjusted by GMT offset. This may mean that we end up on Jan 1 of the next year instead
of this year. Using .startTime() has a symmetrical issue if the GMT offset is negative
giving us the previous year. The current solution is to use the middle of the year.