Fixed Response Time and Process Time hanging in Chart view
Had to use eval in order to construct a mixed type array.
The original code generated a number array that has a string as the first element.
Ferite doesn't allow this. There is a bug in Ferite that makes this possible
using the array literal syntax. So in order to construct such an array
with dynamic size we first construct the string which declares such an array
then eval it.

TODO: Remove the necessity for such an illegal data structure since if Ferite
fixes this bug in the future we can't do this anymore. One option is to use the
following data structure instead:

   [ 'label' => rawRow['x'], 'values' => [0,0,0,0 ...] ]

Another problem: since response time and process time can generate lots of columns
we run out of colors to render them in in chart view.