Fix problem with month May displaying wrong result in drill down
The problem was the english name of the month is used to turn
the name of the month to the month's number representation but
when the user uses the system in Swedish the code was given the swedish name.
This led to May returning 10 instead of 05.

This has been solved by when transling report group outputs then
the original name is also stored. The drill down report can then
make use of the original (english) name of the month.