Fetch errands in drill down from statistics table
The tables workflow_errands and workflow_reporterranddatas are not always completely synced.
There might be good reasons for this (can't think of any right now).
All steps except the last one in the current drill down report is based on the values in workflow_reporterranddatas.
Change so that also the last step (the errands step) also uses this table to make sure the last step shows the same
thing as all the other steps.