Fix problem with channel group being displayed twice in table
In a previous change the untranslated name where saved with a key
that could look like 'group0_original'. This caused another part
of the code to pick this up and think it is a real group that
should be displayed. This happened because the string started
with 'group'.

Fix this by naming the untranslated version 'original_group0'.