Show real total average time for groups in Chart output
The Chart output draws a legend to the right of the chart.
Next to the names of the different groups the output shows the total value.
Before this commit the total value was always calculated by adding all values for that group.
This is not how you calculate a total average time.

Change the key figures total() function to also calculate the total value per group instead of just the total value for the report.

The return value should now look like this.
[ 'total' => 99, 'group0' => [ 'Area A' => 43, 'Area B' => 87' ]