Pass along both the total for x and g to Chart output
Previously the Chart output only showed the real total average value
in the Chart output among in the right side legend.
The total value displayed inside the actual Cart output was calculated
by adding all the values for one row. For a report displaying average values
this isn't the real total value.

Solve this by passing along the total value for both x and g to the data source used by the Chart output.
The Chart output will use the total value for x inside the rendered chart and becaues of this display
the real total average value for the rows.