Fix problem in drill down causing it only display data for 1 area
The problem was that when no area or organisation was specified
the report runs the SQL query with all the system groups
organisations and areas. This is to make sure the drill down
only display data for the current system group.
This caused the JSON output to think that each step should
specify one area and one organisations causing the report to only
display data for one area and one organisation.

Solve the problem by resetting the area and organisation setting
after the report has been run before the JSON output is returned.