FS#4056 - Report table misaligned
This bug is triggered when two groups contain items of the same name. We used a hash (list)
to keep track of rowspan for groups. Unfortunately if more than one group has items of the
same name this confuses the `list` hash. Solved by appending column index to the hashed
value so that the `list` hash can keep track of the items separately.