FS#4241: Translate time group items to user's language
Time group items (example: Monday, December) were being translated
to the user's selected language if the group name only consisted of
one item.

This is not always true because it is possible to build your own
set of time group items.

This fixes that problem by running String.replace() for all known
time group format items and translate them to the right language
if it isn't possible to translate the whole name with a simple I().