FS#4262 - Reports: Deleted errands report showing strange numbers for 'untagged'
Turns out the query was counting errands twice - once when it gets system tagged
(such as deleted, opened etc) and once when it gets tagged.

We do want the system tagged (or completely untagged) errands but not when it
already have client tags. The work-around is very hackish but it seems to work.
We simply move the substr.... clause from the workflow_tags join up one level
to the workflow_errand_tag_links join. But since that table doesn't have tag
names to substr.. we use a subquery to get a list of valid (or in this case
invalid) tag ids