FS-4344-Data format change for reports Average Time for Process,Response
This bug is cause by different formats of time calculation when displaying.
The fix is to change the format of time corresponding to the reports select. For default
reports would format as HH:MM:SS, and only the corresponding values for specific
reports such as days,hours,minutes and seconds would be displayed.