FS#4500 - Fix wrong sort order of months in incoming report
Build report result list would loop through from start date to end date with
an interval of 1 month. This would certaily miss the last date of the range
and being omited from the initial list, and only added at the end when query
result is matched to this build result. At this point the list already is sorted
and adding a new item to the list would cause sort disorder.